requirements.txt
pandas
openpyxl
requests
beautifulsoup4
python-telegram-bot
selenium
webdriver-manager
python-dotenv
xlrd
openpyxl